… 10_gce_docker_guide/figures/ec2-launch-screenshot.png
… 10_gce_docker_guide/figures/launched-ec2-screenshot-1043600.png
… 10_gce_docker_guide/figures/launched-ec2-screenshot.png
… 10_gce_docker_guide/setup/all_zones.txt
… VERSION